moved data vault and registry to core
